+++ src/pykeen/models/multimodal/complex_literal.py
@@ -10,7 +10,7 @@ import torch.nn as nn
 from .base import LiteralModel
 from ...constants import DEFAULT_DROPOUT_HPO_RANGE, DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
 from ...losses import BCEWithLogitsLoss, Loss
-from ...nn.combinations import ComplExLiteralCombination
+from ...nn.combination import ComplExLiteralCombination
 from ...nn.modules import ComplExInteraction, Interaction
 from ...triples import TriplesNumericLiteralsFactory
