+++ src/pykeen/models/multimodal/distmult_literal_gated.py
@@ -8,7 +8,7 @@ import torch.nn as nn
 
 from .base import LiteralModel
 from ...constants import DEFAULT_DROPOUT_HPO_RANGE, DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
-from ...nn.combinations import GatedCombination
+from ...nn.combination import GatedCombination
 from ...nn.modules import DistMultInteraction, Interaction
 from ...triples import TriplesNumericLiteralsFactory
