+++ src/pykeen/nn/representation.py
@@ -18,7 +18,7 @@ from class_resolver.contrib.torch import activation_resolver
 from torch import nn
 from torch.nn import functional
 
-from .combinations import Combination, combination_resolver
+from .combination import Combination, combination_resolver
 from .compositions import CompositionModule, composition_resolver
 from .init import initializer_resolver, uniform_norm_p1_
 from .utils import TransformerEncoder, WikidataCache
