+++ tests/test_nn/test_combination.py
@@ -1,18 +1,18 @@
 """Tests for combination modules."""
 import unittest_templates
 
-import pykeen.nn.combinations
+import pykeen.nn.combination
 from tests import cases
 
 
 class ConcatCombinationTest(cases.CombinationTestCase):
     """Test for concat combination."""
 
-    cls = pykeen.nn.combinations.ConcatCombination
+    cls = pykeen.nn.combination.ConcatCombination
 
 
-class CombinationMetaTestCase(unittest_templates.MetaTestCase[pykeen.nn.combinations.Combination]):
+class CombinationMetaTestCase(unittest_templates.MetaTestCase[pykeen.nn.combination.Combination]):
     """Test for tests for combinations."""
 
-    base_cls = pykeen.nn.combinations.Combination
+    base_cls = pykeen.nn.combination.Combination
     base_test = cases.CombinationTestCase
